don't collect UB instructions for a blockpc without predictions
Previously, we rely on the nullability of UBInstPrecondition to
separate UB instructions from PCs and BlockPCs. However, in cases
where the relevant Phi nodes are not part of the Souper IR,
we can have a null precondition for a BlockPC. Consequently,
we would introduce UB into souper.

This patch fixed this issue by explictly check if we currently
process a PC or BlockPC.